add decorator to reduction to isolate the stored state
Currently custom reducers do not ignore this state rather the reducer morphs the incoming reduction to preserve / clear data store state. Thus we need a way to inject the correct store values to the custom reducer behaviour.

running vs default mode needs to ignore or reflect on state in the custom reducer behaviour this decorate reflects on the running mode reducer state and modifies the reduction store data accordingly to divergent behaviour.